requirements.txt
google-generativeai>=0.3.0
python-dotenv>=0.19.0
pandas>=1.5.0
numpy>=1.21.0
PyPDF2>=3.0.0
python-magic>=0.4.27 